fix: cov/corr respect f32 type
