+++ py-polars/tests/unit/operations/test_statistics.py
@@ -1,4 +1,7 @@
 from datetime import timedelta
+from typing import cast
+
+import pytest
 
 import polars as pl
 from polars.testing import assert_frame_equal
@@ -47,3 +50,61 @@ def test_correlation_cast_supertype() -> None:
     assert df.select(pl.corr("a", "b")).to_dict(as_series=False) == {
         "a": [0.5447047794019219]
     }
+
+
+def test_cov_corr_f32_type() -> None:
+    df = pl.DataFrame({"a": [1, 8, 3], "b": [4, 5, 2]}).select(
+        pl.all().cast(pl.Float32)
+    )
+    assert df.select(pl.cov("a", "b")).dtypes == [pl.Float32]
+    assert df.select(pl.corr("a", "b")).dtypes == [pl.Float32]
+
+
+def test_cov(fruits_cars: pl.DataFrame) -> None:
+    ldf = fruits_cars.lazy()
+    cov_a_b = pl.cov(pl.col("A"), pl.col("B"))
+    cov_ab = pl.cov("A", "B")
+    assert cast(float, ldf.select(cov_a_b).collect().item()) == -2.5
+    assert cast(float, ldf.select(cov_ab).collect().item()) == -2.5
+
+
+def test_std(fruits_cars: pl.DataFrame) -> None:
+    assert fruits_cars.lazy().std().collect()["A"][0] == pytest.approx(
+        1.5811388300841898
+    )
+
+
+def test_var(fruits_cars: pl.DataFrame) -> None:
+    assert fruits_cars.lazy().var().collect()["A"][0] == pytest.approx(2.5)
+
+
+def test_max(fruits_cars: pl.DataFrame) -> None:
+    assert fruits_cars.lazy().max().collect()["A"][0] == 5
+    assert fruits_cars.select(pl.col("A").max())["A"][0] == 5
+
+
+def test_min(fruits_cars: pl.DataFrame) -> None:
+    assert fruits_cars.lazy().min().collect()["A"][0] == 1
+    assert fruits_cars.select(pl.col("A").min())["A"][0] == 1
+
+
+def test_median(fruits_cars: pl.DataFrame) -> None:
+    assert fruits_cars.lazy().median().collect()["A"][0] == 3
+    assert fruits_cars.select(pl.col("A").median())["A"][0] == 3
+
+
+def test_quantile(fruits_cars: pl.DataFrame) -> None:
+    assert fruits_cars.lazy().quantile(0.25, "nearest").collect()["A"][0] == 2
+    assert fruits_cars.select(pl.col("A").quantile(0.25, "nearest"))["A"][0] == 2
+
+    assert fruits_cars.lazy().quantile(0.24, "lower").collect()["A"][0] == 1
+    assert fruits_cars.select(pl.col("A").quantile(0.24, "lower"))["A"][0] == 1
+
+    assert fruits_cars.lazy().quantile(0.26, "higher").collect()["A"][0] == 3
+    assert fruits_cars.select(pl.col("A").quantile(0.26, "higher"))["A"][0] == 3
+
+    assert fruits_cars.lazy().quantile(0.24, "midpoint").collect()["A"][0] == 1.5
+    assert fruits_cars.select(pl.col("A").quantile(0.24, "midpoint"))["A"][0] == 1.5
+
+    assert fruits_cars.lazy().quantile(0.24, "linear").collect()["A"][0] == 1.96
+    assert fruits_cars.select(pl.col("A").quantile(0.24, "linear"))["A"][0] == 1.96
